Re-refactored package names
Also added build and run scripts
