+++ src/com/sontendbox/konsole_combat/client/Main.java
@@ -1,6 +1,6 @@
-package com.sontendbox.client;
+package com.sontendbox.konsole_combat.client;
 
-import com.sontendbox.fight.Controller;
+import com.sontendbox.konsole_combat.Controller;
 
 class Main {
     public static void main(String[] args) {
