+++ test/com/sontendbox/konsole_combat/FIghterTest.java
@@ -1,4 +1,4 @@
-package com.sontendbox.fight;
+package com.sontendbox.konsole_combat;
 
 import static org.junit.Assert.*;
 import org.junit.Before;
